requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
idna==3.10
markdown-it-py==3.0.0
mdurl==0.1.2
Pygments==2.19.1
pytest==8.0.0
pytest-cov==4.1.0
python-dotenv==1.1.0
requests==2.32.3
rich==14.0.0
schedule==1.2.2
shellingham==1.5.4
tinydb==4.8.2
twilio==8.10.0
typer==0.15.2
typing_extensions==4.13.1
urllib3==2.3.0
